build: quiet some "gcc -Wall" compiler warnings
Quiet a number of simple compiler warnings:
- pointers not initialized by ext2fs_get_mem()
- return without value in non-void function
- dereferencing type-punned pointers
- unused variables

+++ lib/ext2fs/crc32c.c
@@ -1117,12 +1117,12 @@ static int test_crc32c(void)
 		be = ext2fs_crc32c_be(t->crc, test_buf + t->start, t->length);
 		if (le != t->crc_le) {
 			printf("Test %d LE fails, %x != %x\n",
-			       (t - test), le, t->crc_le);
+			       (int) (t - test), le, t->crc_le);
 			failures++;
 		}
 		if (be != t->crc_be) {
 			printf("Test %d BE fails, %x != %x\n",
-			       (t - test), be, t->crc_be);
+			       (int) (t - test), be, t->crc_be);
 			failures++;
 		}
 		t++;
